% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cereals}
\alias{cereals}
\title{cereals}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 16 columns.
}
\usage{
cereals
}
\description{
A database containing the qualitative values of different cereals variables
}
\details{
\describe{
\item{species}{barley, durum wheat and soft wheat}
\item{source}{The data comes from different experiences of seed companies or researchers in France and abroad (Denmark, Serbia)}
\item{variable}{The data are classified according to three variables: morphological, phenological and reproductive}
}
}
\keyword{datasets}
