% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legumes}
\alias{legumes}
\title{legumes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75 rows and 26 columns.
}
\usage{
legumes
}
\description{
A database containing the qualitative values of different legumes variables
}
\details{
\describe{
\item{species}{Fababean, pea and lentil}
\item{source}{The data comes from different experiences of seed companies or researchers in France and abroad (Denmark, Serbia)}
\item{variable}{The data are classified according to three variables: morphological, phenological and reproductive}
}
}
\keyword{datasets}
