% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata_quali}
\alias{metadata_quali}
\title{metadata_quali}
\format{
An object of class \code{list} of length 15.
}
\usage{
metadata_quali
}
\description{
a database containing all information on the legume and cereal database. As well as a different sheet for each scale of quantitative variables
}
\details{
\describe{
\item{symbol}{Link with legume and cereal databases}
\item{description}{Definition of the different variables}
\item{sheet}{Each sheet corresponds to a quantitative scale of the variables used}
}
}
\keyword{datasets}
