% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_quali.R
\name{read_metadata_quali}
\alias{read_metadata_quali}
\title{read_metadata_quali}
\usage{
read_metadata_quali()
}
\value{
nothing
}
\description{
read data metadata_quali and allows you to obtain the scales of each variable
}
